refactor(package): 애플리케이션 패키지 구조 개선
