+++ src/main/java/lotto/application/LottoMachine.java
@@ -1,8 +1,10 @@
-package lotto.domain;
+package lotto.application;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.List;
+import lotto.domain.Lotto;
+import lotto.domain.Lottos;
 
 public class LottoMachine {
