+++ src/test/java/lotto/application/LottoMachineTest.java
@@ -1,5 +1,6 @@
-package lotto.domain;
+package lotto.application;
 
+import lotto.domain.Lottos;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
