+++ src/test/java/lotto/view/input/validator/LottoNumberValidatorTest.java
@@ -1,4 +1,4 @@
-package lotto.validator;
+package lotto.view.input.validator;
 
 
 import lotto.exception.common.ErrorMessage;
